# backend/openapi.yaml
openapi: 3.0.0
info:
  title: QQM API
  version: 1.0.0
  description: Client / project / infrastructure management API. Generated from
    the actual implemented routes, controllers, and zod validators
    (backend/src/openapi) — not from any prior design document.
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: /
    description: Same-origin backend
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT obtained from POST /api/auth/login, sent as `Authorization:
        Bearer <token>`."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - message
      required:
        - error
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required:
        - page
        - per_page
        - total
        - total_pages
...
    ScheduleListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleListItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    ScheduleListItem:
      allOf:
        - $ref: "#/components/schemas/Schedule"
        - type: object
          properties:
            is_overdue:
              type: boolean
          required:
            - is_overdue
    ScheduleType:
      type: string
      enum:
        - PM
        - MA
        - other
    ScheduleStatus:
      type: string
      enum:
        - pending
        - in_progress
        - done
        - cancelled
    Schedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          nullable: true
          format: uuid
        server_id:
          type: string
          nullable: true
          format: uuid
        title:
          type: string
        type:
          $ref: "#/components/schemas/ScheduleType"
        scheduled_date:
          type: string
        started_at:
          type: string
          nullable: true
          format: date-time
        completed_at:
          type: string
          nullable: true
          format: date-time
        assigned_to:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ScheduleStatus"
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          nullable: true
          format: date-time
      required:
        - id
        - project_id
        - server_id
        - title
        - type
        - scheduled_date
        - started_at
        - completed_at
        - assigned_to
        - status
        - notes
        - created_at
        - updated_at
        - deleted_at
    ScheduleDetail:
      allOf:
        - $ref: "#/components/schemas/Schedule"
        - type: object
          properties:
            is_overdue:
              type: boolean
            assigned_to_person:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
              required:
                - id
                - name
            project:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
              required:
                - id
                - name
            server:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
              required:
                - id
                - name
          required:
            - is_overdue
            - assigned_to_person
            - project
            - server
...
paths:
...
  /api/schedules:
    get:
      tags:
        - Schedules
      operationId: listSchedules
      summary: List schedules (each item includes a computed is_overdue field)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            description: "Default: 1"
          required: false
          description: "Default: 1"
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            description: "Default: 20, max: 100"
          required: false
          description: "Default: 20, max: 100"
          name: per_page
          in: query
        - schema:
            type: string
            enum:
              - scheduled_date
              - created_at
              - updated_at
            description: "Default: scheduled_date"
          required: false
          description: "Default: scheduled_date"
          name: sort
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            description: "Default: asc"
          required: false
          description: "Default: asc"
          name: order
          in: query
        - schema:
            type: string
            enum:
              - "false"
              - "true"
              - all
            description: "Default: false"
          required: false
          description: "Default: false"
          name: deleted
          in: query
        - schema:
            type: string
            format: uuid
          required: false
          name: project_id
          in: query
        - schema:
            type: string
          required: false
          name: status
          in: query
        - schema:
            type: string
            description: Lower bound on scheduled_date
          required: false
          description: Lower bound on scheduled_date
          name: from
          in: query
        - schema:
            type: string
            description: Upper bound on scheduled_date
          required: false
          description: Upper bound on scheduled_date
          name: to
          in: query
      responses:
        "200":
          description: Paginated schedules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Schedules
      operationId: createSchedule
      summary: Create a schedule (admin or member; status is always 'pending' on create)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: string
                  format: uuid
                server_id:
                  type: string
                  format: uuid
                title:
                  type: string
                  minLength: 1
                type:
                  type: string
                  enum:
                    - PM
                    - MA
                    - other
                scheduled_date:
                  type: string
                assigned_to:
                  type: string
                  format: uuid
                notes:
                  type: string
              required:
                - title
                - type
                - scheduled_date
                - assigned_to
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/schedules/{id}:
    get:
      tags:
        - Schedules
      operationId: getScheduleById
      summary: Get a schedule by id, with assignee/project/server inlined
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Schedules
      operationId: updateSchedule
      summary: Update a schedule's status/notes (admin or member). Status transitions
        are validated by a strict state machine; started_at/completed_at are set
        automatically and cannot be provided directly.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - pending
                    - in_progress
                    - done
                    - cancelled
                notes:
                  type: string
                updated_at:
                  type: string
                  minLength: 1
              required:
                - updated_at
      responses:
        "200":
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Schedules
      operationId: deleteSchedule
      summary: Soft-delete a schedule (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Schedule soft-deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/schedules/{id}/restore:
    post:
      tags:
        - Schedules
      operationId: restoreSchedule
      summary: Restore a soft-deleted schedule (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Schedule restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...